CODERUB: Separate Logic Test For Retrieve All Operation - Student Guardians
